CIV-9198 - removed lint error + moved insetText method
