+++ src/main/common/models/caseProgression/uploadYourDocumentsSectionBuilder.ts
@@ -1,20 +1,7 @@
 import {ClaimSummarySection, ClaimSummaryType} from 'form/models/claimSummarySection';
 import {PageSectionBuilder} from 'common/utils/pageSectionBuilder';
-import {t} from 'i18next';
 export class UploadYourDocumentsSectionBuilder extends PageSectionBuilder {
   _claimSummarySections: ClaimSummarySection[] = [];
-
-  addInsetText(text: string, variables?: unknown) {
-    const insetSection = ({
-      type: ClaimSummaryType.INSET_TEXT,
-      data: {
-        html: t(text),
-        variables: variables,
-      },
-    });
-    this._claimSummarySections.push(insetSection);
-    return this;
-  }
 
   addStartButton(title: string, href: string) {
     const startButtonSection = ({
